+++ Changelog.md
@@ -17,7 +17,11 @@
 ### Misc Changes
 
 - [#584]: Export `EscapeError` from the crate
+- [#581]: Relax requirements for `unsescape_*` set of functions -- their now use
+  `FnMut` instead of `Fn` for `resolve_entity` parameters, like `Iterator::map`
+  from `std`.
 
+[#581]: https://github.com/tafia/quick-xml/pull/581
 [#584]: https://github.com/tafia/quick-xml/pull/584